server: move auth utilities to authUtil
