+++ server/routes/api/torrents.test.ts
@@ -7,7 +7,7 @@ import stream from 'stream';
 import supertest from 'supertest';
 
 import app from '../../app';
-import {getAuthToken} from './auth';
+import {getAuthToken} from '../../util/authUtil';
 import {getTempPath} from '../../models/TemporaryStorage';
 import paths from '../../../shared/config/paths';
